// frontend/src/pages/cadastro/CadastroDePessoaJuridica/PessoaJuridicaForms/types.ts
export interface PessoaJuridica {
  id: string;
  codigo: string;
  razao_social: string;
  nome_fantasia: string;
  cnpj: string;
  inscricao_estadual: string;
  inscricao_municipal: string;
  email: string;
  telefone_comercial: string;
  endereco: string;
  complemento: string;
  bairro: string;
  cidade: string;
  estado: string;
  pais: string;
  cep: string;
  responsavel: string;
  observacoes: string;
}